Separate the model run logic from the repainting timer
We now have two timers running, one running the model
at (as close as possible to) 60 steps/second, and
one repainting the screen using requestAnimFrame, to
repaint at the browser's desired frequency.

See more discussion of why it is good practice to
separate the logic and repainting at
http://www.chandlerprall.com/2012/06/requestanimationframe-is-not-your-logics-friend/

This also allows us to speed up or slow down to model
without needing to change requestAnimFrame, which will
be useful in the next commit where we add a function
for this.