Add setter for step frequency.
We can now call Model.setStepFrequency(x) to make
the model call step() at a frequency of x Hz.

This function can be called at any time, allowing
us to have a speed slider in a model that can be
used by a student to speed up or slow down the model.